restructure Profile
